A preprint version of the research article.
This research article reports on the methodology of constructing a hybrid of empirical and numerical modeling for the first time
and the impact of transient electric-current fluctuations on proton beams in its space-charge-dominated regime. Please note that
if this version needs further updating, it will be reposted in the same place.